network-error-logging/sends-report-on-subdomain-404.https.html

<!DOCTYPE HTML>
<html>
<head>
  <title>
    Test that NEL reports are sent for HTTP errors via an include_subdomains
    policy
  </title>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <script src='./support/fetch.sub.js'></script>
</head>
<body>
  <script>
    promise_test(async t => {
      await clearReportingAndNELConfigurations();
      // Make a request to a resource whose response headers include an
      // include_subdomains NEL policy.
      await fetchResourceWithIncludeSubdomainsPolicy();
      // Make a request to another resource on a subdomain of the above.  This
      // resource doesn't exist, so the server should return a 404.
      await fetchMissingResource('www2');
      // The include_subdomains policy that we just received should cover the
      // second request, so the collector should have received a report about
      // it.
      assert_true(await reportExists({
        url: getURLForMissingResource('www2'),
        type: "network-error",
        report: {
          uri: getURLForMissingResource('www2'),
          sampling_fraction: 1.0,
          status_code: 404,
          type: "http.error",
        },
      }));
    });
  </script>
</body>
</html>
